refactored into leagues.js and leagueController.js
