+++ backend/controllers/leagueController.js
@@ -128,6 +128,23 @@ export const deleteLeague = async (req, res) => {
     });
 };
 
+export const getAddressInfo = async (address) => {
+  const apiKey = process.env.GEOAPIFY; // Use the environment variable
+  const config = {
+    method: 'get',
+    url: `https://api.geoapify.com/v1/geocode/autocomplete?text=${encodeURIComponent(address)}&apiKey=${apiKey}`,
+    headers: {}
+  };
+
+  try {
+    const response = await axios(config);
+    return response.data;
+  } catch (error) {
+    console.error('Error fetching address information:', error);
+    throw error;
+  }
+};
+
 /*
 Sends a request to the captain of the team
 Captain of the team should be in req.params
