+++ backend/routes/leagues.js
@@ -2,7 +2,7 @@ import express from "express";
 
 const router = express.Router();
 
-import { createLeague, updateLeague, getLeagues, getLeague, deleteLeague, getCourts, sendRequestEmail, testroute } from "../controllers/leagueController.js";
+import { createLeague, updateLeague, getLeagues, getLeague, deleteLeague, sendRequestEmail, testroute, getAddressInfo } from "../controllers/leagueController.js";
 
 
 router.route("/sendRequestEmail").get(sendRequestEmail)
@@ -12,6 +12,7 @@ router.route("/:zip").get(getLeague)
 router.route("/updateLeague/:id").patch(updateLeague)
 router.route("/deleteLeague/:id").delete(deleteLeague)  
 router.route("/").get(getLeagues)
-router.get("/courts", getCourts);
+router.route("/address").get(getAddressInfo);
+
 
 export default router;
